+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.1 (unreleased)
 
 - Add L2 MTU module
+- Add interface type module
 
 ## 0.2.0